services/atomic/route/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
googlemaps==4.10.0